SUPPORTED: document what CentOS means
The operating system we support is CentOS Linux: the operating system
with periodic releases that is roughly equivalent to Red Hat Enterprise
Linux.  We don't support CentOS Stream because it's a rolling release
distro, which are difficult to support due to frequent changes, and
consequently the only one of those we support is Debian unstable.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>